add test for resources widget
